src/xiangji/requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.7
cmake==3.27.2
comm==0.1.4
contourpy==1.1.0
cupy-cuda11x==12.2.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
diffusers==0.14.0
docker-pycreds==0.4.0
einops==0.6.1
executing==1.2.0
fastrlock==0.8.2
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
importlib-metadata==6.8.0
install==1.3.5
ipdb==0.13.13
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
kornia==0.7.0
lightning-utilities==0.9.0
lit==16.0.6
lpips==0.1.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pie-torch==0.0.1
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
Pygments==2.16.1
pyparsing==3.1.1
pyre-extensions==0.0.29
python-dateutil==2.8.2
pytorch-fid==0.1.1
pytorch-lightning==1.4.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scipy==1.10.1
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
sympy==1.12
taming-transformers==0.0.1
taming-transformers-rom1504==0.0.6
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
test-tube==0.7.5
tokenizers==0.12.1
tomli==2.0.1
torch==2.0.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.25.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.14.0
wcwidth==0.2.6
Werkzeug==2.3.7
xformers==0.0.20
yarl==1.9.2
zipp==3.16.2
